anlikhaber/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
openai==1.3.5
pillow==10.0.1
instagrapi==1.18.4
beautifulsoup4==4.12.2
lxml==4.9.3
apscheduler==3.10.4
gunicorn==21.2.0
python-dateutil==2.8.2
werkzeug==2.3.7
jinja2==3.1.2
wtforms==3.0.1
email-validator==2.0.0
flask-migrate==4.0.5 